+++ packages/test/src/container/selectors.ts
@@ -3,34 +3,29 @@
  */
 
 export const CONTAINER_SELECTORS = {
-  // 主要按钮选择器
-  showBtn: '[data-testid="show-container-btn"]',
-  hideBtn: '[data-testid="hide-container-btn"]',
+  // 主要控制按钮选择器
+  toggleShowBtn: '[data-testid="toggle-show-btn"]',
   toggleFullscreenBtn: '[data-testid="toggle-fullscreen-btn"]',
 
   // 容器相关选择器
   container: '[data-testid="test-container"]',
-  containerStatus: '[data-testid="container-status"]',
-  fullscreenStatus: '[data-testid="fullscreen-status"]',
+  containerDraggingBar: '.tr-container__dragging-bar',
+  containerHeader: '.tr-container__header',
+  containerHeaderOperations: '.tr-container__header-operations',
+  containerFooter: '.tr-container__footer',
 
-  // 日志相关选择器
-  actionLog: '[data-testid="action-log"]',
-  clearLogsBtn: '[data-testid="clear-logs-btn"]',
+  // 容器内置按钮选择器
+  containerCloseBtn: '.tr-container__header-operations button:last-child',
+  containerFullscreenBtn: '.tr-container__header-operations button:nth-last-child(2)',
 
-  // 输入相关选择器
-  demoInput: '[data-testid="demo-input"]',
-  demoInputValue: '[data-testid="demo-input-value"]',
-
-  // 容器内部元素选择器
+  // 插槽内容选择器
   containerTitle: '[data-testid="container-title"]',
   containerContent: '[data-testid="container-content"]',
-  containerFooter: '[data-testid="container-footer"]',
-  containerDraggingBar: '.tr-container__dragging-bar',
-  containerHeaderOperations: '.tr-container__header-operations',
-
-  // 自定义操作按钮选择器
+  containerFooterContent: '[data-testid="container-footer"]',
   customOperationBtn: '[data-testid="custom-operation-btn"]',
-  footerActionBtn: '[data-testid="footer-action-btn"]',
+
+  // 测试用输入框
+  demoInput: '[data-testid="demo-input"]',
 } as const
 
 /**
